v3.0 (#36)

+++ entrypoint.sh
@@ -9,15 +9,15 @@ fi
 
 # Keep FlexGet updated
 echo "Checking for FlexGet updates ..."
-pip3 install --upgrade --quiet pip
-pip3 install --upgrade --quiet flexget~=${VERSION}
+pip${PYTHON_VERSION/2/} install --upgrade --quiet pip
+pip${PYTHON_VERSION/2/} install --upgrade --quiet flexget~=${FLEXGET_VERSION}
 
 # Default FlexGet config
-if [[ ! -f /config/config.yml ]]; then
+if [[ ! -f /config/config.yml && -f /flexget.yml ]]; then
     cp /flexget.yml /config/config.yml
 fi
 
 # Set web UI password
-flexget web passwd ${PASSWORD}
+flexget web passwd ${FLEXGET_PASSWORD}
 
 exec "$@"
